requirements.txt
alembic==1.12.0
annotated-types==0.6.0
asgiref==3.8.1
asyncmy==0.2.10
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
coverage==7.4.0
cryptography==41.0.4
dnspython==2.4.2
dominate==2.8.0
exceptiongroup==1.1.3
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
freezegun==1.4.0
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jpholiday==0.1.8
Mako==1.2.4
MarkupSafe==2.1.3
MonthDelta==0.9.1
numpy==1.26.0
packaging==23.2
pandas==2.2.3
pluggy==1.5.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-freezegun==0.4.2
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.2
requests==2.31.0
six==1.16.0
SkPy==0.10.6
soupsieve==2.5
SQLAlchemy==2.0.21
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2024.2
urllib3==1.26.6
visitor==0.1.3
Werkzeug==2.3.8
WTForms==3.0.1
wtforms-html5==0.6.1
zipp==3.17.0
